moved remaining classes to entiy item module
