+++ felt-entity-item-api/src/main/java/io/github/feltmc/feltapi/api/entityitem/EntityCustomItem.java
@@ -1,13 +1,10 @@
-package io.github.feltmc.feltapi.api.item.extensions;
+package io.github.feltmc.feltapi.api.entityitem;
 
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-
-import javax.annotation.Nullable;
 
 public interface EntityCustomItem {
     private Item self()
