+++ felt-entity-item-api/src/main/resources/entity-item.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "io.github.feltmc.feltapi.mixin.item",
+  "package": "io.github.feltmc.feltapi.mixin.entityitem",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "ItemEntityMixin"
