--- felt-item-api/build.gradle
@@ -1,14 +1,0 @@
-archivesBaseName = "felt-item-api"
-version = getSubprojectVersion(project)
-
-repositories {
-    maven { url = "https://mvn.devos.one/snapshots/" }
-}
-
-dependencies {
-    modImplementation("io.github.fabricators_of_create:Porting-Lib:${rootProject.port_lib_version}-beta+1.18.2-stable"){
-        exclude(group: "com.github.LlamaLad7")
-    }
-    implementation(include("com.github.LlamaLad7:MixinExtras:0.1.1"))
-    annotationProcessor("com.github.LlamaLad7:MixinExtras:0.1.1")
-}
