--- felt-item-api/src/main/java/io/github/feltmc/feltapi/impl/item/ItemPrelaunch.java
@@ -1,11 +1,0 @@
-package io.github.feltmc.feltapi.impl.item;
-
-import com.llamalad7.mixinextras.MixinExtrasBootstrap;
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
-
-public class ItemPrelaunch implements PreLaunchEntrypoint {
-    @Override
-    public void onPreLaunch() {
-        MixinExtrasBootstrap.init();
-    }
-}
